requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-ratelimit==4.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
django-extensions==3.2.3

# PostGIS / spatial
# GDAL must match your system GDAL version. Install separately:
#   pip install GDAL=="$(gdal-config --version)"
# Or via conda: conda install -c conda-forge gdal

# HTTP clients
requests==2.32.3
urllib3==2.2.3

# PDF generation
reportlab==4.2.2
Pillow>=10.0.0
qrcode[pil]>=7.4

# Accident data (Czech Police monthly RAR/XLS)
rarfile==4.2
pyproj==3.7.1

# Utilities
gunicorn==22.0.0
whitenoise==6.7.0
